pyproject.toml
[build-system]
requires = ["uv_build>=0.11.0,<0.12"]
build-backend = "uv_build"

[project]
name = "snowflake-cortex-agents"
version = "0.1.0"
description = "Python SDK for Snowflake Cortex Agents and Analyst"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    {name = "Chuliang Xiao", email = "ChuliangX@gmail.com"}
]
keywords = ["snowflake", "cortex", "agents", "analyst", "ai", "llm", "rest"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Database",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dotenv = [
    "python-dotenv>=1.0.0",
]
charts = [
    "altair>=5.0.0",
    "pandas>=1.0.0",
]
dev = [
    "pre-commit>=4.0.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.7.0",
    "ty>=0.0.26",
]
docs = [
    "mkdocs>=1.5.0,<2.0",
    "mkdocs-material>=9.0.0",
    "mkdocstrings[python]>=0.24.0",
]
all = [
    "snowflake-cortex-agents[dotenv,charts,dev,docs]",
]

[project.urls]
Homepage = "https://github.com/ChuliangXiao/snowflake-cortex-agents"
Documentation = "https://github.com/ChuliangXiao/snowflake-cortex-agents/blob/main/README.md"
Repository = "https://github.com/ChuliangXiao/snowflake-cortex-agents"
Issues = "https://github.com/ChuliangXiao/snowflake-cortex-agents/issues"

[tool.uv.build-backend]
module-name = "cortex_agents"
module-root = ""

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E203",  # whitespace before ':'
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ty.analysis]
allowed-unresolved-imports = ["IPython.**", "altair", "httpx", "dotenv"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=cortex_agents",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]

[tool.coverage.run]
source = ["cortex_agents"]
omit = [
    "*/tests/*",
    "*/examples/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
